LastCommit

+++ app/Http/Controllers/Backend/BrandController.php
@@ -5,8 +5,10 @@ namespace App\Http\Controllers\Backend;
 use App\DataTables\BrandDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Brand;
+use App\Models\Product;
 use App\Traits\ImageUploadTrait;
 use Illuminate\Http\Request;
+use PhpOffice\PhpSpreadsheet\Cell\IValueBinder;
 use Str;
 
 class BrandController extends Controller
@@ -52,7 +54,6 @@ class BrandController extends Controller
 
         toastr('Created Successfully!', 'success');
         return redirect()->route('admin.brand.index');
-
     }
 
     /**
@@ -78,7 +79,7 @@ class BrandController extends Controller
     public function update(Request $request, string $id)
     {
         $request->validate([
-            'logo' => ['image', 'nullabel', 'max:2000'],
+            'logo' => ['image', 'max:2000'],
             'name' => ['required', 'max:200'],
             'is_featured' => ['required'],
             'status' => ['required']
@@ -86,7 +87,6 @@ class BrandController extends Controller
 
         $brand = Brand::findOrFail($id);
 
-         /** Handle file update */
         $logoPath = $this->updateImage($request, 'logo', 'uploads', $brand->logo);
 
         $brand->logo = empty(!$logoPath) ? $logoPath : $brand->logo;
@@ -106,8 +106,12 @@ class BrandController extends Controller
     public function destroy(string $id)
     {
         $brand = Brand::findOrFail($id);
-        $this -> deleteImage($brand->logo);
+        if(Product::where('brand_id', $brand->id)->count() > 0){
+            return response(['status' => 'error', 'message' => 'This brand have products you can\'t delete it.']);
+        }
+        $this->deleteImage($brand->logo);
         $brand->delete();
+
         return response(['status' => 'success', 'message' => 'Deleted Successfully!']);
     }
