+++ app/Http/Controllers/Backend/ChildCategoryController.php
@@ -6,6 +6,8 @@ use App\DataTables\ChildCategoryDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Category;
 use App\Models\ChildCategory;
+use App\Models\HomePageSetting;
+use App\Models\Product;
 use App\Models\SubCategory;
 use Illuminate\Http\Request;
 use Str;
@@ -26,8 +28,7 @@ class ChildCategoryController extends Controller
     public function create()
     {
         $categories = Category::all();
-        $subCategories = SubCategory::all();
-        return view('admin.child-category.create', compact('categories', 'subCategories'));
+        return view('admin.child-category.create', compact('categories'));
     }
 
     /**
@@ -46,19 +47,19 @@ class ChildCategoryController extends Controller
     {
         $request->validate([
             'category' => ['required'],
-            'sub_category'=>['required'],
+            'sub_category' => ['required'],
             'name' => ['required', 'max:200', 'unique:child_categories,name'],
             'status' => ['required']
         ]);
 
         $childCategory = new ChildCategory();
-        
-        $childCategory -> category_id = $request->category;
-        $childCategory -> sub_category_id = $request->sub_category;
-        $childCategory -> name = $request->name;
-        $childCategory -> slug = Str::slug($request->name);
-        $childCategory -> status = $request->status;
-        $childCategory -> save();
+
+        $childCategory->category_id = $request->category;
+        $childCategory->sub_category_id = $request->sub_category;
+        $childCategory->name = $request->name;
+        $childCategory->slug = Str::slug($request->name);
+        $childCategory->status = $request->status;
+        $childCategory->save();
 
         toastr('Created Successfully!', 'success');
 
@@ -71,7 +72,7 @@ class ChildCategoryController extends Controller
     public function show(string $id)
     {
         //
-    }       
+    }
 
     /**
      * Show the form for editing the specified resource.
@@ -83,7 +84,7 @@ class ChildCategoryController extends Controller
         $subCategories = SubCategory::where('category_id', $childCategory->category_id)->get();
 
         return view('admin.child-category.edit', compact('categories', 'childCategory', 'subCategories'));
-    }   
+    }
 
     /**
      * Update the specified resource in storage.
@@ -92,21 +93,21 @@ class ChildCategoryController extends Controller
     {
         $request->validate([
             'category' => ['required'],
-            'sub_category'=>['required'],
-            'name' => ['required', 'max:200', 'unique:child_categories,name,'. $id],
+            'sub_category' => ['required'],
+            'name' => ['required', 'max:200', 'unique:child_categories,name,'.$id],
             'status' => ['required']
         ]);
 
-        $childCategory =  ChildCategory::findOrFail($id);
-        
-        $childCategory -> category_id = $request->category;
-        $childCategory -> sub_category_id = $request->sub_category;
-        $childCategory -> name = $request->name;
-        $childCategory -> slug = Str::slug($request->name);
-        $childCategory -> status = $request->status;
-        $childCategory -> save();
+        $childCategory = ChildCategory::findOrFail($id);
 
-        toastr('Updated Successfully!', 'success');
+        $childCategory->category_id = $request->category;
+        $childCategory->sub_category_id = $request->sub_category;
+        $childCategory->name = $request->name;
+        $childCategory->slug = Str::slug($request->name);
+        $childCategory->status = $request->status;
+        $childCategory->save();
+
+        toastr('Update Successfully!', 'success');
 
         return redirect()->route('admin.child-category.index');
     }
@@ -117,6 +118,19 @@ class ChildCategoryController extends Controller
     public function destroy(string $id)
     {
         $childCategory = ChildCategory::findOrFail($id);
+        if(Product::where('child_category_id', $childCategory->id)->count() > 0){
+            return response(['status' => 'error', 'message' => 'This item contain relation can\'t delete it.']);
+        }
+        $homeSettings = HomePageSetting::all();
+
+        foreach($homeSettings as $item){
+            $array = json_decode($item->value, true);
+            $collection = collect($array);
+            if($collection->contains('child_category', $childCategory->id)){
+                return response(['status' => 'error', 'message' => 'This item contain relation can\'t delete it.']);
+            }
+        }
+
         $childCategory->delete();
 
         return response(['status' => 'success', 'message' => 'Deleted Successfully!']);
@@ -124,9 +138,9 @@ class ChildCategoryController extends Controller
 
     public function changeStatus(Request $request)
     {
-    $childCategory = ChildCategory::findOrFail($request->id);
-        $childCategory->status = $request->status == 'true' ? 1 : 0;
-        $childCategory->save();
+        $category = ChildCategory::findOrFail($request->id);
+        $category->status = $request->status == 'true' ? 1 : 0;
+        $category->save();
 
         return response(['message' => 'Status has been updated!']);
     }
